chore(beta releases): do not push tag or changelog in beta releases
just bump build publish

So that we keep the full changelog for the good release

+++ scripts/release-beta.sh
@@ -27,33 +27,43 @@ npm install
 
 currentVersion=`cat package.json | json version`
 
-# header
-printf "\n\nBeta release: current version is $currentVersion"
-printf "\nBeta release: a changelog will be generated only if a fix/feat/performance/breaking token is found in git log"
-printf "\nBeta release: you must choose a new ve.rs.ion (semver)"
-printf "\nBeta release: press q to exit the next screen\n\n"
+# if not on a beta version already
+if [[ $currentVersion != *"-beta"* || $CHOOSE_BUMP ]]
+then
+  # header
+  printf "\n\nBeta release: current version is $currentVersion"
+  printf "\nBeta release: a changelog will be generated only if a fix/feat/performance/breaking token is found in git log"
+  printf "\nBeta release: you must choose a new ve.rs.ion (semver)"
+  printf "\nBeta release: press q to exit the next screen\n\n"
 
-# preview changelog
-read -p "=> Beta release: press [ENTER] to view changes since latest version.."
+  # preview changelog
+  read -p "=> Beta release: press [ENTER] to view changes since latest version.."
 
-# nd (markdown renderer in terminal)
-# has some output going on to stderr because of node v4, no big deal
-conventional-changelog -p angular | nd | less -r 2>/dev/null
+  # nd (markdown renderer in terminal)
+  # has some output going on to stderr because of node v4, no big deal
+  conventional-changelog -p angular | nd | less -r 2>/dev/null
 
-# choose and bump new version
-# printf "\n\nBeta release: Please enter the new chosen version > "
-printf "\n=> Beta release: please type the new chosen version (-beta will be automatically appended) > "
-read -e newVersion
-newVersion="$newVersion-beta"
+  # choose and bump new version
+  # printf "\n\nBeta release: Please enter the new chosen version > "
+  printf "\n=> Beta release: please type the new chosen version (-beta will be automatically appended) > "
+  read -e newVersion
+  newVersion="$newVersion-beta.0"
+else
+  # when already on a beta version, let's bump the beta-.X
+  newVersion=`semver ${currentVersion} -i prerelease -preid beta`
+
+  read -p "=> Beta release: The new version will be ${newVersion},
+  press [ENTER] to continue, CTRL+C to abort.
+  You can force specifying a different bump by using CHOOSE_BUMP=1 npm run release-beta"
+fi
+
+echo $newVersion
+exit
+
 VERSION=$newVersion babel-node ./scripts/bump-package-version.js
 
 # build new version
 NODE_ENV=production VERSION=$newVersion npm run build
-
-# update changelog
-printf "\n\nBeta release: update changelog"
-changelog=`conventional-changelog -p angular`
-conventional-changelog -p angular -i CHANGELOG.md -w
 
 # regenerate readme TOC
 printf "\n\nBeta release: generate README.mc TOC"
@@ -67,12 +77,10 @@ npm run jsdoc:widget
 commitMessage="v$newVersion\n\n$changelog"
 git add src/lib/version.js npm-shrinkwrap.json package.json CHANGELOG.md README.md docs/_includes/widget-jsdoc
 printf "$commitMessage" | git commit --file -
-git tag "v$newVersion"
 
 printf "\n\nBeta release: almost done, check everything in another terminal tab.\n"
 read -p "=> Beta release: when ready, press [ENTER] to push to github and publish the package to the beta channel"
 
 printf "\n\nBeta release: push to github, publish on npm"
 git push origin develop
-git push origin --tags
 npm publish --tag beta
